refactor alloc drain to make intent more clear (#25731)
While working on #25726, I found a method in the drainer code that records
creates a map of job IDs to allocations.

At first glance this looks like a bug because it effectively de-duplicates the
allocations per job. But the consumer of the map is only concerned with jobs,
not allocations, and simply reads the job off the allocation. Refactor this to
make it obvious we're looking at the job.

Ref: https://github.com/hashicorp/nomad/pull/25726