refactor procstats name and comment
The `linuxProcStats` type in the executor's `procstats` package is misnamed as a
result of a couple rounds of refactoring. It's used by all task executors, not
just Linux. Rename this and move a comment about how Windows processes are
listed so that the comment is closer to where the logic is implemented.