network: Support IPv6 only interfaces in CNI
In previous versions, Nomad supported IPv6 only interfaces by simply
setting the V6 value in `Address`. This worked until the move to
`AddressV6` in the networking CNI code, making interfaces that only use
IPv6 unable to be used.

This commit tries to come back to that behaviour, where Address could
also be a V6.

The reason we are not renaming to AddressIPv4 is to be retrocompatible
with stored allocation networks on places like BoltDB.