scheduler: refactor cluster reconciler to make its functions explicitly return results
For readability and testing, it would be better for reconciler logic to
be expressed in terms of explicit, composable functions instead of
methods of the allocReconciler object. This way there are no hidden
mutations and we are always certain exactly what is being passed as
arguments and what is being returned.