tls: Reset server TLS authenticator when TLS config reloaded
The Nomad server uses an authenticator backend for RPC handling
which includes TLS verification. This verification setting is
configured based on the servers TLS configuration object and is
built when a new server is constructed.

The bug occurs when a servers TLS configuration is reloaded which
can change the desired TLS verification handling. In this case,
the authenticator is not updated, meaning the RPC mTLS verification
is not modified, even if the configuration indicates it should.

This change adds a new function on the authenticator to allow
updating its TLS verification rule. This new function is called
when a servers TLS configuration is reloaded.

backport of commit a3e096b0c98e3e76bcf153e0eaae4ef7d5184e42 (#26126)

Co-authored-by: James Rasell <jrasell@users.noreply.github.com>