cli: Do not always add global DNS name to certificate DNS names
No matter the passed region identifier, the CLI was always adding
"<role>.global.nomad" to the certificate DNS names. This is not
what we expect and has been removed.

While here, the long deprecated cluster-region flag has been
removed. This removal only impacts CLI functionality, so is safe
to do.

---------

Co-authored-by: James Rasell <jrasell@users.noreply.github.com>