fsm: Add identity signing key handling for node status updates. (#26139)
When a node heartbeats, the RPC handler will optionally generate
an identity to return to the caller. The identity key ID will be
stored in the node object, so we have tracking of keys in use.

The state store has been updated to handle node status update
requests that include a signing key ID. Rather than add another
parameter into the function signature, the FSM function now takes
the entire request object.