client: Handle identities from servers and use for RPC auth.
Nomad servers, if upgraded, can return node identities as part of
the register and update/heartbeat response objects. The Nomad
client will now handle this and store it as appropriate within its
memory and statedb.

The client will now use any stored identity for RPC authentication
with a fallback to the secretID. Ths supports upgrades paths where
the Nomad clients are updated before the Nomad servers.